bug: increased accuracy to double in counting intensities

+++ BpmVideoProcessor.h
@@ -26,7 +26,7 @@ class BpmVideoProcessor {
     vector<Mat> blurredForMask; // Blurred video in set level
     vector<Mat> temporalSpatial; // Temporal spatial
 
-    vector<int> intensities; // Intensities of blurred frames
+    vector<double> intensities; // Intensities of blurred frames
 
     // Computed bpm
     int bpm;
